changed net code from double to float for speed reasons.

+++ mimir-test/TestMatrix.cpp
@@ -8,7 +8,7 @@ REGISTER_TEST(TestMatrix)
 
 using mimir::models::Matrix;
 
-static bool valarraysEqual(const std::valarray<double> &left, const std::valarray<double> &right)
+static bool valarraysEqual(const std::valarray<float> &left, const std::valarray<float> &right)
 {
     return (left == right).min() == true;
 }
@@ -73,7 +73,7 @@ void TestMatrix::testMultiply()
     QVERIFY(valarraysEqual({158, 184, 210},result[1]));
     QVERIFY(valarraysEqual({246, 288, 330},result[2]));
 
-    Matrix m3{std::vector<std::valarray<double>>{{1, 2}}};
+    Matrix m3{std::vector<std::valarray<float>>{{1, 2}}};
     QVERIFY_EXCEPTION_THROWN(m1*m3, std::logic_error);
 }
 
@@ -86,18 +86,18 @@ void TestMatrix::testOtherMultiplications()
             {7, 8, 9}
         }
     };
-    std::vector<double> vector{2, 2, 2};
+    std::vector<float> vector{2, 2, 2};
     m1 *= vector;
-    QVERIFY(m1 == Matrix(std::vector<std::valarray<double>>{{12},{30},{48}}));
-    m1 = Matrix(std::vector<std::valarray<double>>{{1},{3},{9}});
-    std::valarray<double> array{2, 2, 2};
+    QVERIFY(m1 == Matrix(std::vector<std::valarray<float>>{{12},{30},{48}}));
+    m1 = Matrix(std::vector<std::valarray<float>>{{1},{3},{9}});
+    std::valarray<float> array{2, 2, 2};
     m1 *= array;
     QVERIFY(m1 == Matrix({{2,2,2},{6,6,6},{18,18,18}}));
     m1 *= .5;
     QVERIFY(m1 == Matrix({{1,1,1},{3,3,3},{9,9,9}}));
     m1 = Matrix({{.9, .1, .2}, {.9, .1, .2}, {.9, .1, .2}});
-    Matrix m2 = m1 * std::vector<double>{.2, .2, .2};
-    QVERIFY(m2 == Matrix(std::vector<std::valarray<double>>{{.24},{.24},{.24}}));
+    Matrix m2 = m1 * std::vector<float>{.2, .2, .2};
+    QVERIFY(m2 == Matrix(std::vector<std::valarray<float>>{{.24},{.24},{.24}}));
 }
 
 void TestMatrix::testDiff()
@@ -119,7 +119,7 @@ void TestMatrix::testDiff()
                 {1, 1, 1, 1, 1},
                 {1, 1, 1, 1, 1},
                 {1, 1, 1, 1, 1}});
-    m1 -= Matrix(std::vector<double>{.1, .1, .1, .1, .1});
+    m1 -= Matrix(std::vector<float>{.1, .1, .1, .1, .1});
     QVERIFY(m1 == Matrix({
                              { .9, .9, .9, .9, .9},
                              { .9, .9, .9, .9, .9},
@@ -132,7 +132,7 @@ void TestMatrix::testDiff()
                 {1, 1, 1, 1, 1},
                 {1, 1, 1, 1, 1},
                 {1, 1, 1, 1, 1}});
-    m1 -= std::vector<double>{.1, .1, .1, .1, .1};
+    m1 -= std::vector<float>{.1, .1, .1, .1, .1};
     QVERIFY(m1 == Matrix({
                              { .9, .9, .9, .9, .9},
                              { .9, .9, .9, .9, .9},
@@ -144,15 +144,15 @@ void TestMatrix::testDiff()
                     {1, 2, 3},
                     {4, 5, 6}
                   });
-    m1 -= std::valarray<double>{1, 2, 3};
+    m1 -= std::valarray<float>{1, 2, 3};
     QVERIFY(m1 == Matrix ({{0, 0, 0},
                            {3, 3, 3}}));
 }
 
 void TestMatrix::testCallbackConstructor()
 {
-    double vals[2][3] = {{1, 2, 3}, {4, 5, 6}};
-    auto setter = [&vals](size_t row, size_t col) -> double { return vals[row][col];};
+    float vals[2][3] = {{1, 2, 3}, {4, 5, 6}};
+    auto setter = [&vals](size_t row, size_t col) -> float { return vals[row][col];};
     Matrix cbAssigned(2, 3, setter);
     QVERIFY(cbAssigned == Matrix({{1, 2, 3}, {4, 5, 6}}));
 }
