+++ mimir/models/activation/AbstractMSELossActivator.cpp
@@ -6,14 +6,14 @@ namespace mimir {
 namespace models {
 namespace activation {
 
-double AbstractMSELossActivator::error(const mimir::models::TrainerValueHelper &item) const noexcept
+float AbstractMSELossActivator::error(const mimir::models::TrainerValueHelper &item) const noexcept
 {
     auto diff = item.hypothesis() - item.expectation();
     diff *= diff;
-    return  std::accumulate(diff.begin(), diff.end(), 0.)/static_cast<double>(diff.size());
+    return  std::accumulate(diff.begin(), diff.end(), 0.)/static_cast<float>(diff.size());
 }
 
-std::vector<double> AbstractMSELossActivator::lossDerivative(const std::vector<double> &hypothesis, const std::vector<double> &expectations) const noexcept
+std::vector<float> AbstractMSELossActivator::lossDerivative(const std::vector<float> &hypothesis, const std::vector<float> &expectations) const noexcept
 {
     return (hypothesis - expectations) * derivative(hypothesis);
 }
