+++ mimir/models/activation/ActivationInterface.h
@@ -16,13 +16,13 @@ class ActivationInterface
 public:
     virtual ~ActivationInterface() = default;
     virtual std::string name() const = 0;
-    virtual void activate(std::vector<double> &) const noexcept = 0;
-    virtual std::vector<double> derivative(std::vector<double> const &hypothesis) const noexcept = 0;
-    virtual std::vector<double> lossDerivative(std::vector<double> const &hypothesis, std::vector<double> const &expectations) const noexcept = 0;
-    virtual double error(mimir::models::TrainerValueHelper const &values) const noexcept = 0;
+    virtual void activate(std::vector<float> &) const noexcept = 0;
+    virtual std::vector<float> derivative(std::vector<float> const &hypothesis) const noexcept = 0;
+    virtual std::vector<float> lossDerivative(std::vector<float> const &hypothesis, std::vector<float> const &expectations) const noexcept = 0;
+    virtual float error(mimir::models::TrainerValueHelper const &values) const noexcept = 0;
 protected:
     template<typename Op>
-    inline void apply(std::vector<double> &vec, Op operation) const noexcept {
+    inline void apply(std::vector<float> &vec, Op operation) const noexcept {
         std::transform(vec.begin(), vec.end(), vec.begin(), operation);
     }
 };
