+++ mimir/models/activation/RectifiedLinear.cpp
@@ -10,16 +10,16 @@ std::string RectifiedLinear::name() const
     return "rectifiedLinear";
 }
 
-void RectifiedLinear::activate(std::vector<double> &v) const noexcept
+void RectifiedLinear::activate(std::vector<float> &v) const noexcept
 {
-    std::transform(v.begin(), v.end(), v.begin(), [] (double v) -> double { return v > 0. ? v : .1 * v; });
+    std::transform(v.begin(), v.end(), v.begin(), [] (float v) -> float { return v > 0. ? v : .1 * v; });
 }
 
-std::vector<double> RectifiedLinear::derivative(const std::vector<double> &hypothesis) const noexcept
+std::vector<float> RectifiedLinear::derivative(const std::vector<float> &hypothesis) const noexcept
 {
     auto derivative = hypothesis;
     // let's make that leaky relu
-    apply(derivative, [](double x) { return x > 0. ? 1. : 0.1; });
+    apply(derivative, [](float x) { return x > 0. ? 1. : 0.1; });
     return derivative;
 }
