+++ mimir/models/activation/Sigmoid.h
@@ -11,8 +11,8 @@ class Sigmoid : public AbstractMSELossActivator
 {
 public:
     std::string name() const override;
-    void activate(std::vector<double> &) const noexcept override;
-    std::vector<double> derivative(std::vector<double> const &hypothesis) const noexcept override;
+    void activate(std::vector<float> &) const noexcept override;
+    std::vector<float> derivative(std::vector<float> const &hypothesis) const noexcept override;
 protected:
 };
